+++ servers/registry.js
@@ -31,7 +31,7 @@ app.use(router.allowedMethods());
 app.use(smartNotFound);
 
 app.on("error", (err, ctx) => {
-  process.emit("catch-error", err, ctx);
+  process.emit("systemError", err, ctx);
   console.log(err);
   console.log(ctx);
 });